fix krazy qminmax hit (portability issue)
git-svn-id: svn+ssh://svn.kdab.com/home/SVN-klaralv/products/kdchart/trunk/kdchart@40128 c66179b2-ff0d-0410-8aef-1b1a91e65cfa

+++ examples/Zoom/Keyboard/zoomwidget.cpp
@@ -112,7 +112,7 @@ void ZoomWidget::keyPressEvent( QKeyEvent* e )
             e->ignore();
     }
     if( dZoom || dX || dY ){
-        const double factor = qMax( 1.0, coordinatePlane()->zoomFactorX() + dZoom );
+        const double factor = qMax( double(1.0), coordinatePlane()->zoomFactorX() + dZoom );
         const qreal x = coordinatePlane()->zoomCenter().x() + dX;
         const qreal y = coordinatePlane()->zoomCenter().y() + dY;
         coordinatePlane()->setZoomFactorX( factor );
